+++ app/src/main/java/romanow/abc/ess2/android/rendering/view2/DesktopGUIData.java
@@ -57,12 +57,12 @@ public class DesktopGUIData extends View2BaseDesktop {
         context.getMain().main().popupAndLog(ss);
         }
     @Override
-    public void putValue(int vv) throws UniException {
+    public void putValue(long vv) throws UniException {
         Meta2Register register = getRegister();
         if (((Meta2GUIData)getElement()).isByteSize()){
             textField.setText(""+(byte)vv);
             return;
-        }
+            }
         int type = register.getFormat();
         if (type==Values.FloatValue)
             textField.setText(""+Float.intBitsToFloat((int)vv));
