+++ app/src/main/java/romanow/abc/ess2/android/rendering/view2/DesktopGUIDataLabel.java
@@ -34,7 +34,7 @@ public class DesktopGUIDataLabel extends View2BaseDesktop {
         context.getMain().main().popupAndLog(ss);
         }
     @Override
-    public void putValue(int vv) throws UniException {
+    public void putValue(long vv) throws UniException {
         Meta2Register register = getRegister();
         if (((Meta2GUIData)getElement()).isByteSize()){
             label.setText(""+(byte)vv);
