+++ app/src/main/java/romanow/abc/ess2/android/rendering/view2/DesktopGUILevelIndicator.java
@@ -55,11 +55,11 @@ public class DesktopGUILevelIndicator extends View2BaseDesktop {
         setInfoClick(textField);
         }
     @Override
-    public void putValue(Meta2Register register, int xx, int idx) {
+    public void putValue(Meta2Register register, long xx, int idx) {
         limits[idx] = register.regValueToFloat(getUnitIdx(),(int)xx);
         }
     @Override
-    public void putValue(int xx) throws UniException {
+    public void putValue(long xx) throws UniException {
         Meta2DataRegister register = (Meta2DataRegister) getRegister();
         if (((Meta2GUILevelIndicator)getElement()).isByteSize()){
             value = (byte)xx;
